+++ src/st/no1/unk_35E20.c
@@ -1861,289 +1861,3 @@ INCLUDE_ASM("st/no1/nonmatchings/unk_35E20", func_us_801BB4C0);
 INCLUDE_ASM("st/no1/nonmatchings/unk_35E20", func_us_801BB5E0);
 
 INCLUDE_ASM("st/no1/nonmatchings/unk_35E20", func_us_801BB7B8);
-
-INCLUDE_ASM("st/no1/nonmatchings/unk_35E20", func_us_801BB984);
-
-extern u16 D_us_801809F8[];
-extern RECT D_us_801B47B0;
-extern s32 D_psp_08C62AA4;
-
-void func_us_801BBD90(Entity* self) {
-    s16 dx, dy;
-    s16 var9a;
-    s32 primIndex;
-    s32 dx2;
-    Entity* tempEntity;
-    DRAWENV drawEnv;
-    DR_ENV* dr_env;
-    s16 dAngle;
-    s16 posX, posY;
-    s16 angle;
-    s32 i;
-    s16 magnitude;
-    s16 temp;
-    Primitive* prim;
-    s32 dy2;
-    RECT unused;
-    u8 padding[4];
-
-    unused = D_us_801B47B0;
-    tempEntity = self - 1;
-    self->posX.i.hi = tempEntity->posX.i.hi;
-    self->posY.i.hi = tempEntity->posY.i.hi - 0x22;
-    magnitude = 8;
-    switch (self->step) {
-    case 0:
-        InitializeEntity(D_us_801809F8);
-        self->animCurFrame = 0;
-        break;
-
-    case 1:
-        primIndex = g_api.func_800EDB58(PRIM_GT4, 0x20);
-        if (primIndex != -1) {
-            self->flags |= FLAG_HAS_PRIMS;
-            self->primIndex = primIndex;
-            prim = &g_PrimBuf[primIndex];
-            dr_env = g_api.func_800EDB08((POLY_GT4*)prim);
-            if (dr_env == NULL) {
-                DestroyEntity(self);
-                return;
-            }
-            prim->type = PRIM_ENV;
-            drawEnv = g_CurrentBuffer->draw;
-            drawEnv.isbg = 1;
-            drawEnv.r0 = 0;
-            drawEnv.g0 = 0;
-            drawEnv.b0 = 0;
-            drawEnv.clip.x = 0;
-            drawEnv.clip.w = 0x20;
-            drawEnv.clip.y = 0x100;
-            drawEnv.clip.h = 0x100;
-            drawEnv.ofs[0] = 0;
-            drawEnv.ofs[1] = 0x100;
-            SetDrawEnv(dr_env, &drawEnv);
-            prim->priority = 0x11F;
-            prim->drawMode = DRAW_HIDE;
-            prim = prim->next;
-            dr_env = g_api.func_800EDB08((POLY_GT4*)prim);
-            if (dr_env == NULL) {
-                DestroyEntity(self);
-                return;
-            }
-            prim->type = PRIM_ENV;
-            prim->priority = 0x121;
-            prim->drawMode = DRAW_UNK_800 | DRAW_HIDE;
-            prim = prim->next;
-
-            self->ext.et_801BBD90.unk7C = prim;
-#ifdef VERSION_PSP
-            prim->tpage = 0x8010;
-#else
-            prim->tpage = 0x10;
-#endif
-            prim->clut = 0x2F;
-            prim->u0 = 0;
-            prim->u1 = prim->u0 + 0x10;
-            prim->u2 = prim->u0;
-            prim->u3 = prim->u1;
-            prim->v0 = 0;
-            prim->v1 = prim->v0;
-            prim->v2 = 0xFF;
-            prim->v3 = prim->v2;
-            prim->priority = 0x69;
-            prim = prim->next;
-
-            while (prim != NULL) {
-                prim->tpage = 0x10E;
-                prim->clut = 0x2F;
-                prim->u0 = 0x3E;
-                prim->u1 = prim->u0 + 2;
-                prim->u2 = prim->u0;
-                prim->u3 = prim->u1;
-                prim->v0 = 0xC0;
-                prim->v1 = prim->v0;
-                prim->v2 = prim->v0 + 0x10;
-                prim->v3 = prim->v2;
-                prim->priority = 0x120;
-                prim->drawMode = DRAW_DEFAULT;
-                prim = prim->next;
-            }
-        } else {
-            DestroyEntity(self);
-            return;
-        }
-
-        self->step++;
-        break;
-
-    case 2:
-        prim = &g_PrimBuf[self->primIndex];
-        prim->drawMode = DRAW_DEFAULT;
-        prim = prim->next;
-        prim->drawMode = DRAW_UNK_800;
-        prim = prim->next;
-        prim = self->ext.et_801BBD90.unk7C;
-        prim = prim->next;
-        posX = 0;
-        posY = 0;
-        while (prim != NULL) {
-            prim->tpage = 0x10E;
-            prim->x0 = prim->x2 = posX;
-            prim->y0 = prim->y1 = posY;
-            prim->x1 = prim->x3 = prim->x2 + 2;
-            posY = prim->y2 = prim->y3 = prim->y0 + 0x10;
-            prim = prim->next;
-        }
-        self->step_s = 0;
-        self->step++;
-        break;
-
-    case 3:
-        if (!self->step_s) {
-            prim = &g_PrimBuf[self->primIndex];
-            prim->drawMode = DRAW_HIDE;
-            prim = prim->next;
-            prim->drawMode = DRAW_HIDE;
-            prim = prim->next;
-            prim = self->ext.et_801BBD90.unk7C;
-            prim = prim->next;
-            while (prim != NULL) {
-                prim->v0 = prim->v1 = 0xC0;
-                prim->v2 = prim->v3 = 0xC8;
-                prim = prim->next;
-            }
-            self->step_s++;
-#ifdef VERSION_PSP
-            D_psp_08C62AA4 = 1;
-#endif
-        }
-        if ((self->posX.i.hi & 0x100) || (self->posY.i.hi < 0)) {
-            prim = self->ext.et_801BBD90.unk7C;
-            while (prim != NULL) {
-                prim->drawMode = DRAW_HIDE;
-                prim = prim->next;
-            }
-        } else {
-            i = 0;
-            dAngle = 0;
-            posX = self->posX.i.hi;
-            posY = self->posY.i.hi;
-            posX -= 4;
-            if (posY > 0xF0) {
-                posY = (self->posY.i.hi & 0xF) + 0xF0;
-            }
-            tempEntity = self - 1;
-            if (tempEntity->ext.et_801BBD90.unk87) {
-                if (tempEntity->velocityY < 0) {
-                    self->ext.et_801BBD90.unk8C += 0x60;
-                } else {
-                    if (self->ext.et_801BBD90.unk8C > 0) {
-                        self->ext.et_801BBD90.unk8C -= 0x30;
-                    } else {
-                        self->ext.et_801BBD90.unk8C += 0x30;
-                    }
-                }
-            } else if (self->ext.et_801BBD90.unk8C) {
-                if (self->ext.et_801BBD90.unk8C > 0) {
-                    self->ext.et_801BBD90.unk8C -= 0x30;
-                } else {
-                    self->ext.et_801BBD90.unk8C += 0x30;
-                }
-            }
-            prim = self->ext.et_801BBD90.unk7C;
-            if (!self->ext.et_801BBD90.unk8C) {
-                prim->drawMode = DRAW_UNK02;
-                prim->priority = 0x69;
-                prim->x0 = prim->x2 = posX;
-                prim->y2 = prim->y3 = posY;
-                prim->x1 = prim->x0 + 0x10;
-                prim->x3 = prim->x2 + 0x10;
-                prim->y0 = prim->y1 = prim->y2 - 0xF0;
-                prim->v2 = prim->v3 = 0xFF;
-                temp = g_Tilemap.scrollY.i.hi + prim->y0;
-                if (temp < 0x300) {
-                    prim->y0 = 0x300 - g_Tilemap.scrollY.i.hi;
-                    prim->y1 = prim->y0;
-                }
-                prim->v0 = prim->v1 = prim->v2 - (prim->y2 - prim->y0);
-                prim = prim->next;
-                while (prim != NULL) {
-                    prim->drawMode = DRAW_HIDE;
-                    prim = prim->next;
-                }
-            } else {
-                prim->drawMode = DRAW_HIDE;
-                prim = prim->next;
-                angle = self->ext.et_801BBD90.unk8C;
-                angle &= 0xFFF;
-                while (prim != NULL) {
-#ifdef VERSION_PSP
-                    prim->tpage = 0x10;
-                    prim->u0 = prim->u2 = 0;
-                    prim->u1 = prim->u3 = 4;
-                    prim->v0 = prim->v1 = 0;
-                    prim->v2 = prim->v3 = 0x10;
-#else
-                    prim->tpage = 0xE;
-                    prim->u0 = -6;
-                    prim->u3 = prim->u1 = -2;
-                    prim->u2 = -6;
-#endif
-                    prim->priority = 0x69;
-                    prim->x2 = posX;
-                    prim->y2 = posY;
-                    dx = -(magnitude * rsin(angle)) >> 0xC;
-                    dy = -(magnitude * rcos(angle)) >> 0xC;
-                    prim->x0 = posX + dx;
-                    prim->y0 = posY + dy;
-                    prim->x1 = prim->x0 + 4;
-                    prim->x3 = prim->x2 + 4;
-                    prim->y1 = prim->y0;
-                    prim->y3 = prim->y2;
-                    if (i == 0) {
-                        i = 4;
-                        dx2 = self->posX.i.hi - 4 - prim->x0;
-                        dy2 = 0x10;
-                        var9a = ratan2(-dx2, dy2);
-                        dAngle = var9a - angle;
-                        if (dAngle > 0x800) {
-                            dAngle = 0x1000 - dAngle;
-                        }
-                        if (dAngle < -0x800) {
-                            dAngle = dAngle + 0x1000;
-                        }
-                        dAngle /= 4;
-                    }
-                    angle += dAngle;
-                    angle &= 0xFFF;
-                    posX = prim->x0;
-                    posY = prim->y0;
-                    prim->drawMode = DRAW_UNK02;
-                    if (prim->y2 < -0x10) {
-                        break;
-                    }
-                    temp = g_Tilemap.scrollY.i.hi + prim->y0;
-                    if (temp < 0x300) {
-                        prim->y0 = 0x300 - g_Tilemap.scrollY.i.hi;
-                        prim->y1 = prim->y0;
-                        break;
-                    }
-                    i--;
-                    prim = prim->next;
-                }
-
-                while (prim != NULL) {
-                    prim->drawMode = DRAW_HIDE;
-                    prim = prim->next;
-                }
-            }
-        }
-        break;
-    }
-}
-
-INCLUDE_ASM("st/no1/nonmatchings/unk_35E20", func_us_801BC598);
-
-INCLUDE_ASM("st/no1/nonmatchings/unk_35E20", func_us_801BC9A8);
-
-INCLUDE_RODATA("st/no1/nonmatchings/unk_35E20", D_us_801B47B0);
